chore: mockRaceResult 위치 변경
